test(progress-view): merge the two StreamTabs suites into one per-module suite
StreamTabsStatus and StreamTabsExpandChevron both exercised the same
StreamTabs component with duplicated scaffolding (StreamTabInfo builder,
settleChildRender, jsdom setup) and a drifted file name. All 16 tests
move verbatim into StreamTabs.vitest.ts sharing one builder (agent
category as a parameter), one render-settle helper, and one passthrough
mount that the status-list mount now delegates to.
