+++ src/test-kernel/progressView/StreamTabs.vitest.ts
@@ -20,19 +20,40 @@ import {
   useLitComponentTestDom,
 } from '../settings/litComponentTestUtils';
 
+useLitComponentTestDom(
+  () => import('@progressView/frontend/components/StreamTabs'),
+);
+
 interface StyledElementConstructor extends CustomElementConstructor {
   readonly elementStyles: readonly (
     CSSStyleSheet | { readonly cssText: string }
   )[];
 }
 
-function stream(name: string): StreamTabInfo {
+function stream(
+  name: string,
+  agentCategory: AgentCategory = AgentCategory.ToolUse,
+): StreamTabInfo {
   return {
     identity: { kind: 'agent', agent: name },
     name,
     label: name,
-    agentCategory: AgentCategory.ToolUse,
+    agentCategory,
     creationTimestamp: 1,
+  };
+}
+
+function makeBashChild(
+  name: string,
+  parentStreamId: string,
+  creationTimestamp: number,
+): StreamTabInfo {
+  return {
+    identity: { kind: 'process', tool: 'bash' },
+    name,
+    label: 'bash',
+    parentStreamId,
+    creationTimestamp,
   };
 }
 
@@ -55,7 +76,13 @@ function settleChildRender(): Promise<unknown> {
   return new Promise((resolve) => setTimeout(resolve, 0));
 }
 
-async function mountTabs(
+async function mountTabs(props: Partial<StreamTabs>): Promise<StreamTabs> {
+  const element = await mountComponent<StreamTabs>('stream-tabs', props);
+  await settleChildRender();
+  return element;
+}
+
+async function mountTabsWithStatuses(
   streams: StreamTabInfo[],
   statuses: readonly StreamLifecycleStatus[],
   activeStreamId: string | null = null,
@@ -66,20 +93,10 @@ async function mountTabs(
       streamState(statuses[index] ?? STREAM_PHASE.RUNNING),
     ]),
   );
-  const tabs = await mountComponent<StreamTabs>('stream-tabs', {
-    streams,
-    streamStates,
-    activeStreamId,
-  });
-  await settleChildRender();
-  return tabs;
+  return mountTabs({ streams, streamStates, activeStreamId });
 }
 
 describe('stream-tab lifecycle distinction', () => {
-  useLitComponentTestDom(
-    () => import('@progressView/frontend/components/StreamTabs'),
-  );
-
   it('shows canonical lifecycle glyphs without redundant visible labels', async () => {
     const streams = [
       stream('running'),
@@ -89,7 +106,7 @@ describe('stream-tab lifecycle distinction', () => {
       stream('cancelled'),
       stream('ready'),
     ];
-    const tabs = await mountTabs(streams, [
+    const tabs = await mountTabsWithStatuses(streams, [
       STREAM_PHASE.RUNNING,
       STREAM_PHASE.WAITING,
       STREAM_PHASE.COMPLETED,
@@ -159,7 +176,7 @@ describe('stream-tab lifecycle distinction', () => {
 
   it('keeps an unknown lifecycle value visible as its defensive fallback', async () => {
     const item = stream('unknown');
-    const tabs = await mountComponent<StreamTabs>('stream-tabs', {
+    const tabs = await mountTabs({
       streams: [item],
       streamStates: new Map([
         [
@@ -171,7 +188,6 @@ describe('stream-tab lifecycle distinction', () => {
         ],
       ]),
     });
-    await settleChildRender();
     const row = tabs.shadowRoot?.querySelector('stream-tab');
 
     expect(
@@ -212,7 +228,7 @@ describe('stream-tab lifecycle distinction', () => {
       ],
       ['approval', streamState(STREAM_PHASE.RUNNING)],
     ]);
-    const tabs = await mountComponent<StreamTabs>('stream-tabs', {
+    const tabs = await mountTabs({
       streams,
       streamStates,
       pendingApprovalStreamIds: new Set(['approval']),
@@ -221,7 +237,6 @@ describe('stream-tab lifecycle distinction', () => {
       ]),
       compact: true,
     });
-    await settleChildRender();
     const rows = [...(tabs.shadowRoot?.querySelectorAll('stream-tab') ?? [])];
 
     expect(
@@ -280,7 +295,7 @@ describe('stream-tab lifecycle distinction', () => {
   });
 
   it('surfaces a descendant approval on the compact parent row', async () => {
-    const tabs = await mountComponent<StreamTabs>('stream-tabs', {
+    const tabs = await mountTabs({
       streams: [stream('parent')],
       streamStates: new Map([['parent', streamState(STREAM_PHASE.RUNNING)]]),
       pendingApprovalStreamIds: new Set(['child']),
@@ -289,7 +304,6 @@ describe('stream-tab lifecycle distinction', () => {
       ]),
       compact: true,
     });
-    await settleChildRender();
     const row = tabs.shadowRoot?.querySelector('stream-tab');
 
     expect(
@@ -311,7 +325,7 @@ describe('stream-tab lifecycle distinction', () => {
   });
 
   it('keeps selection primary while retaining the accessible lifecycle label', async () => {
-    const tabs = await mountTabs(
+    const tabs = await mountTabsWithStatuses(
       [stream('selected-running'), stream('selected-stopped')],
       [STREAM_PHASE.RUNNING, STREAM_PHASE.CANCELLED],
       'selected-running',
@@ -339,7 +353,7 @@ describe('stream-tab lifecycle distinction', () => {
   });
 
   it('preserves long-title truncation, status indicators, semantic tokens, and forced-color overrides', async () => {
-    const tabs = await mountTabs(
+    const tabs = await mountTabsWithStatuses(
       [
         stream(
           'a very long execution label that must yield space to lifecycle status',
@@ -390,5 +404,263 @@ describe('stream-tab lifecycle distinction', () => {
         ?.querySelector('#stream-tab-status')
         ?.getAttribute('aria-label'),
     ).toBe('Completed');
+  });
+});
+
+describe('stream-tab expand chevron', () => {
+  // The child-stream expand/collapse chevron uses the shared `wa-button`
+  // action-icon pattern, and StreamTabs.handleTabClick delegates off
+  // `[data-stream][data-action]`, so the element must carry both attributes.
+  it('renders the expand toggle as a wa-button carrying the delegated-click contract', async () => {
+    const tabs = await mountTabs({
+      streams: [stream('parent', AgentCategory.Workflow)],
+      childStreamsByParent: new Map([
+        ['parent', [stream('child', AgentCategory.Workflow)]],
+      ]),
+    });
+    const parentTab = tabs.shadowRoot?.querySelector('stream-tab');
+    expect(parentTab).toBeTruthy();
+
+    const expandButton = parentTab?.shadowRoot?.querySelector('.tab-expand');
+    expect(expandButton).toBeTruthy();
+    expect(expandButton?.tagName).toBe('WA-BUTTON');
+    expect(expandButton?.getAttribute('data-stream')).toBe('parent');
+    expect(expandButton?.getAttribute('data-action')).toBe('toggle-children');
+    expect(expandButton?.getAttribute('aria-expanded')).toBe('false');
+    expect(expandButton?.getAttribute('aria-label')).toBe('1 background task');
+    expect(
+      parentTab?.shadowRoot
+        ?.querySelector('wa-tooltip[for="stream-tab-expand-button"]')
+        ?.textContent?.trim(),
+    ).toBe('1 background task');
+  });
+
+  it('expands ancestors when a child stream is the active session', async () => {
+    const tabs = await mountTabs({
+      streams: [stream('parent', AgentCategory.Workflow)],
+      childStreamsByParent: new Map([
+        ['parent', [stream('child', AgentCategory.Workflow)]],
+      ]),
+    });
+    const parentTab = tabs.shadowRoot?.querySelector('stream-tab');
+    const expandButton = parentTab?.shadowRoot?.querySelector('.tab-expand');
+    expect(expandButton?.getAttribute('aria-expanded')).toBe('false');
+
+    tabs.activeStreamId = 'child';
+    await tabs.updateComplete;
+    await settleChildRender();
+
+    expect(
+      parentTab?.shadowRoot
+        ?.querySelector('.tab-expand')
+        ?.getAttribute('aria-expanded'),
+    ).toBe('true');
+    expect(tabs.shadowRoot?.querySelectorAll('stream-tab')).toHaveLength(2);
+  });
+
+  it('identifies an unlabeled stream by name in the select aria-label', async () => {
+    const tabs = await mountTabs({
+      streams: [
+        { ...stream('unlabeled-stream', AgentCategory.Workflow), label: '' },
+      ],
+    });
+
+    const ariaLabel = tabs.shadowRoot
+      ?.querySelector('stream-tab')
+      ?.shadowRoot?.querySelector('#stream-tab-select-button')
+      ?.getAttribute('aria-label');
+    expect(ariaLabel).toContain('unlabeled-stream');
+  });
+
+  it('renders no session footer', async () => {
+    const tabs = await mountTabs({
+      streams: [stream('session', AgentCategory.Workflow)],
+    });
+
+    expect(tabs.shadowRoot?.querySelector('.stream-list-footer')).toBeNull();
+    expect(tabs.shadowRoot?.querySelector('stream-tab')).toBeTruthy();
+  });
+
+  it('anchors the general hint to the title, not an ancestor of specific hints', async () => {
+    const tabs = await mountTabs({
+      streams: [
+        {
+          ...stream('remote', AgentCategory.Workflow),
+          isRemote: true,
+          worktree: {
+            workingDirectory: '/tmp/texra/remote',
+            branch: 'remote',
+          },
+        },
+      ],
+    });
+
+    const tab = tabs.shadowRoot?.querySelector('stream-tab');
+    const shadow = tab?.shadowRoot;
+    const title = shadow?.querySelector('#stream-tab-title');
+    const select = shadow?.querySelector('#stream-tab-select-button');
+
+    expect(
+      shadow?.querySelector('wa-tooltip[for="stream-tab-select-button"]'),
+    ).toBeNull();
+    expect(
+      shadow?.querySelector('wa-tooltip[for="stream-tab-title"]'),
+    ).toBeNull();
+    expect(select?.getAttribute('aria-label')).toContain('Status:');
+    expect(
+      title?.contains(shadow?.querySelector('#stream-tab-kind') ?? null),
+    ).toBe(false);
+    expect(
+      title?.contains(shadow?.querySelector('#stream-tab-remote') ?? null),
+    ).toBe(false);
+    expect(
+      title?.contains(shadow?.querySelector('worktree-chip') ?? null),
+    ).toBe(false);
+
+    tabs.compact = true;
+    tabs.childStreamsByParent = new Map([
+      ['remote', [stream('remote-child', AgentCategory.Workflow)]],
+    ]);
+    await tabs.updateComplete;
+    await settleChildRender();
+
+    const compactShadow =
+      tabs.shadowRoot?.querySelector('stream-tab')?.shadowRoot;
+    expect(
+      compactShadow
+        ?.querySelector('#stream-tab-title')
+        ?.contains(compactShadow.querySelector('#stream-tab-compact-children')),
+    ).toBe(false);
+  });
+
+  it('renders workflow scripts as orchestration streams without a model', async () => {
+    const tabs = await mountTabs({
+      streams: [
+        {
+          identity: {
+            kind: 'multiAgentWorkflow',
+            workflowName: 'repo-cleanup-readonly-pilot-2026-07-24',
+          },
+          name: 'workflow-script#abc123',
+          label: 'repo-cleanup-readonly-pilot-2026-07-24',
+          agentCategory: AgentCategory.Workflow,
+          creationTimestamp: 1,
+        },
+      ],
+    });
+
+    const tab = tabs.shadowRoot?.querySelector('stream-tab');
+    const model = tab?.shadowRoot?.querySelector('.model');
+    const kindIcon = tab?.shadowRoot?.querySelector('.stream-kind') as
+      (Element & { name?: string }) | null;
+
+    expect(model?.textContent?.trim()).toBe('');
+    expect(kindIcon?.name).toBe('list-ul');
+    expect(kindIcon?.hasAttribute('title')).toBe(false);
+    expect(
+      tab?.shadowRoot
+        ?.querySelector('wa-tooltip[for="stream-tab-kind"]')
+        ?.textContent?.trim(),
+    ).toBe('Multi-Agent Workflow');
+    expect(tab?.shadowRoot?.querySelector('.agent-name')).toBeNull();
+  });
+
+  it('surfaces the agent name inline on the metadata line when the title is a summary', async () => {
+    const tabs = await mountTabs({
+      streams: [
+        {
+          ...stream('engineer', AgentCategory.Workflow),
+          agentCategory: AgentCategory.ToolUse,
+          description: 'Removing Supabase migrations from remote...',
+          model: 'grok45',
+          modelLabel: 'Grok 4.5',
+        },
+      ],
+    });
+
+    const shadow = tabs.shadowRoot?.querySelector('stream-tab')?.shadowRoot;
+    expect(shadow?.querySelector('#stream-tab-title')?.textContent).toContain(
+      'Removing Supabase migrations from remote...',
+    );
+    expect(shadow?.querySelector('.model')?.textContent?.trim()).toBe(
+      'Grok 4.5',
+    );
+    expect(shadow?.querySelector('.agent-name')?.textContent?.trim()).toBe(
+      'engineer',
+    );
+  });
+
+  it('keeps the opaque stream id out of row text but in its accessible name', async () => {
+    const tabs = await mountTabs({
+      streams: [
+        {
+          ...stream('review', AgentCategory.Workflow),
+          name: 'review#a4c8939992cf',
+          agentCategory: AgentCategory.ToolUse,
+        },
+      ],
+    });
+
+    const shadow = tabs.shadowRoot?.querySelector('stream-tab')?.shadowRoot;
+    // Visible text carries the name only — never the hex suffix.
+    expect(
+      shadow?.querySelector('#stream-tab-title')?.textContent,
+    ).not.toContain('a4c8939992cf');
+    // The row summary stays aria-only by design (#9168) — no visual tooltip
+    // spanning the row — but the id must still be in it, since that is what
+    // separates two parallel runs of the same agent.
+    expect(
+      shadow
+        ?.querySelector('#stream-tab-select-button')
+        ?.getAttribute('aria-label'),
+    ).toContain('review#a4c8939992cf');
+  });
+
+  it('omits the inline agent name when the title already is that same identity', async () => {
+    const tabs = await mountTabs({
+      streams: [
+        {
+          ...stream('engineer', AgentCategory.Workflow),
+          agentCategory: AgentCategory.ToolUse,
+          // No `description`: the title falls back to the identity name
+          // itself, so repeating it on the metadata line would just echo
+          // the title back.
+        },
+      ],
+    });
+
+    const shadow = tabs.shadowRoot?.querySelector('stream-tab')?.shadowRoot;
+    expect(shadow?.querySelector('#stream-tab-title')?.textContent).toContain(
+      'engineer',
+    );
+    expect(shadow?.querySelector('.agent-name')).toBeNull();
+  });
+
+  it('hides finished process children from the Sessions tree', async () => {
+    const parent = {
+      ...stream('engineer', AgentCategory.Workflow),
+      agentCategory: AgentCategory.ToolUse,
+    };
+    const liveBash = makeBashChild('bash@tool#live', 'engineer', 2);
+    const doneBash = makeBashChild('bash@tool#done', 'engineer', 3);
+    const tabs = await mountTabs({
+      streams: [parent],
+      childStreamsByParent: new Map([['engineer', [liveBash, doneBash]]]),
+      streamStates: new Map([
+        ['engineer', { status: 'running', lastTimestamp: 1 } as never],
+        ['bash@tool#live', { status: 'running', lastTimestamp: 2 } as never],
+        ['bash@tool#done', { status: 'completed', lastTimestamp: 3 } as never],
+      ]),
+    });
+
+    const childTabs = [
+      ...(tabs.shadowRoot?.querySelectorAll('.child-streams stream-tab') ?? []),
+    ];
+    expect(childTabs).toHaveLength(1);
+    expect(
+      childTabs[0]?.shadowRoot
+        ?.querySelector('#stream-tab-title')
+        ?.textContent?.trim(),
+    ).toBe('bash');
   });
 });
